Don't expect stopOffload to succeed if interface doesn't exist.
The AdditionalStopsWithInitReturnFalse test inits offload and
then checks that stopOffload returns true. However, the HAL is
allowed to return false to stopOffload if no upstream was ever
successfully programmed.

To ensure this test doesn't fail in the common case that the
interface is not up, don't expect stopOffload to return true in
this case.

Bug: 67439856
Test: VtsHalTetheroffloadControlV1_0TargetTest passes with and without mobile data connection
Change-Id: I245f9e3e4376a7799572a3b3967e2bf6c52b5e4d
(cherry picked from commit 3189d9de81c937d495482fa3dacc196878cfcda1)