Added CellIdentity struct in types.hal
This change was made because the new scan API needs to include the human
readable operator name other than PLMN in the scan results.
In order to let teh operator name used by other structures and APIS, we added
the operator name alphaLong and alphaShort, which were actually included
in the OperatorInfo, into the CellIdentity.

Bug: 63984327
Test: Basic telephony sanity
Change-Id: I498b7f190b739c92975fe797e1c74e96d90ee823