Replace OUT with OUT_DIR
OUT is only defined if some functions in envsetup.sh are run, which is
not the case on the build servers. I'm looking at removing the
environment variable in local builds to keep things consistent.

In this case, OUT_DIR was actually expected, not OUT, which is
equivalent to PRODUCT_OUT.

Test: none
Change-Id: I1e5e9f40727104716212d696927d1a32d7a74fab