fwk comp mat: deprecate health@1.0
By removing 1.0 from current.xml, health@1.0 is officially
considered "deprecated" but not "removed". It is still
supported for upgrading devices.

On upgrading devices, fwk/storaged->healthd->1.0 HAL still
exist, so this entry is kept on 26/27/legacy.xml.

On launch devices, device is required to have health 2.0 HAL,
hence the only code path is fwk/storaged->2.0 HAL. healthd
and 1.0 HAL is not used.

Test: builds (assemble_vintf passes)
Bug: 65990106
Change-Id: I4deb80074fe8e5a399e01e86bffcc69c272e748b