Revert old matrices back to DR / MR1.
legacy/26/27.xml ("old matrices") should not be changed once
published.

The requirements / HAL usage information are now encompassed in
multiple matrices (instead of a single compatibility_matrix.V.xml).
For example, the usage of HALs on Shipping FCM Version 26 is:
* Required HALs in 26.xml
* Using all required + optional HALs in 26.xml/27.xml/...

In particular,

* Starting at 26.xml (diff legacy 26), gatekeeper is required.
* Starting at 27.xml (diff 26 27),
   - broadcastradio@1.0 => 1.0-1
   - introduce cas@1.0
   - introduce oemlock@1.0
   - power@1.0 => 1.0-1
   - radio@1.0 => 1.0-1
   - introduce tetheroffload.config@1.0
   - introduce tetheroffload.control@1.0
   - thermal@1.0 => 1.0-1
   - usb@1.0 => 1.0-1
   - vibrator@1.0 => 1.0-1
   - wifi@1.0 => 1.0-1
   - introduce weaver@1.0
   - introduce wifi.offload@1.0

health@2.0 and configstore@1.1 is not specified
in legacy/26/27.xml, but framework can still use them, since they
are introduced in current.xml.

Test: boot test

Bug: 69861676
Change-Id: I897df6952245a8ce33e1c63264c3970b0483c2f7