Deprecate nfc@1.0 for P.
nfc@1.0 requires a conf file (libnfc-nci.conf) in order to work.
Since this file has no schema, it is a system<->vendor violation
and must be placed on the system image, however this makes the
system image non-generic.

nfc@1.1 contains this configuration in the interface, and so it
can be used entirely without creating a Treble violation.

Test: P-launching devices have nfc@1.1
Bug: 72080121
Change-Id: I7a99eb7cd5e577be0c70b4c5124f28fa471717e6