VtsHalMediaOmx should test functionality only
The tests insisted on presence of codecs as well as functionality, but
the list it reads is a superset of what may be on the device.  Modified so
that if a codec is not present, we ignore it rather than failing the test.

The Vts tests that check OMX Store are being augmented to ensure any
advertised codecs are indeed present.

Bug: 80029024
Bug: 80530946
Test: vts run VtsHalMediaOmxV1_0Host
Change-Id: Icd6d049c5c0d31f47ef3dbda52d6927e38bd894d