Wifi: Set wlan.driver.status to 'ok' when ready
Currently, the wlan.driver.status property is set to 'ok' once the
driver is loaded. This is not accurate, cause having the driver loaded
does not necessarily means it is ready for operation.

This commit sets the driver to 'ok' state after it is checked for
readiness. This implementation of the check is device specific.

Bug: 111924712
Test: Manual test

Change-Id: I074d819e9acdfdd6f033d85cdd71c0bc3984eb9d
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>