Revert "Fixed VTS failure in no SIM case"
Since the EMERGENCY-type APN is a must set in the radio
setupDataCall request for modem to perform Emergency call,
and the given VTS test case does not set that emergency bit,
I think modem should treat the request as a normal call
request, and should not return "NONE" for no-sim.

This reverts commit 431eb118f84de14c94ec18f887197944ec17e712.

Reason for revert: after further discussion, NONE
is not acceptable given the test case.

Bug: 109767888
Change-Id: I3d1cc96120d53a9be0ae5059c26b091bf82dc352