[VTS][RTT] Fix CreateRtt test case: use STA instead of AP
The CreateRtt test case verifies that an RTT controller can be
created (once at least one other interface is created - an AP
in the original test case).

However, recent code change clarified the actual behavior which ties
the RTT controller to the existence (real or potential) of a STA
interface (not an AP interface). This caused test failures with
the original AP mechanism in non-DBS mode (in DBS mode an AP+STA
could exist and so the test passes).

Fix the test by using a STA as the control.

Bug: 118439379
Test: vts-tradefed run commandAndExit vts --module WifiChipHidlTest on
      both DBS and non-DBG devices.

Change-Id: I5a1bfc9a3b54c480ba7a514f90bc48f1fbe7af7d