Update patch set 1
Patch Set 1:

> > Well. On galaxy s6 it address "no audio call (sound) issue".

Maybe it'd be good to add this and other bits discussed here to commit message?

 > On hlte there IS sound, garbled, but still. The BluetoothSCOTest
 > app is even stating "Supports use of SCO for off call: YES"
 > 
 > So I beleive it has more to do with buffers or codecs than this
 > commit.

My feeling as well on hammerhead, some audio HAL changes between N and O, but it's traumatic to even look in there at this point…

> So that other device are not affected, (if at all this commit breaks other device)

For what it's worth, this change seems benign based on your explanation, and no additional HFP breakage on hammerhead with it.

Patch-set: 1