Replace network type with network capabilities in IAgnssRil.hal
The NetworkType enumerator in IAGnssRil.hal is mapped to the
transport type constants (TYPE_MOBILE, TYPE_WIFI, etc.) in the
framework class ConnectivityManager. These constants are deprecated
and their use should be replaced with network capabilities instead.

The GNSS HAL implementation should make decisions based on the
reported capabilities of a network and not on the type of the network.

Bug: 117768826
Test: Tested with code instrumentation to verify that the
      @2.0::AGnssRil.hal updateNetwork_2_0() method implementation
      is called with correct arguments.
Test: Verified @1.0::AGnssRil.hal backward compatibility on a
      Pixel 3 device.
Change-Id: I8f7f8ca8f5254815702b6d5a89dcb4e53a289cfe