DO NOT MERGE: Audio VTS: Wait after stream close
Due to asynchronous nature of the destruction of server-side objects
it is required to flush IPC messages to the server and wait to avoid
flakiness due to an attempt to open the stream while it's still not
closed on the server side.

This patch is specific for Android P release.

Test: atest VtsHalAudioV4_0TargetTest
Bug: 118655804
Change-Id: I1a5ec28bce9802ec654c139153ec4aa6786474e5